[Refactor]: Refactor StreamService.
