+++ grpcapi/streamservice/DalalStreamService.go
@@ -12,8 +12,8 @@ import (
 	"google.golang.org/grpc/codes"
 
 	"github.com/delta/dalal-street-server/datastreams"
-	"github.com/delta/dalal-street-server/proto_build"
-	"github.com/delta/dalal-street-server/proto_build/datastreams"
+	pb "github.com/delta/dalal-street-server/proto_build"
+	datastreams_pb "github.com/delta/dalal-street-server/proto_build/datastreams"
 	"github.com/delta/dalal-street-server/session"
 
 	"github.com/delta/dalal-street-server/utils"
@@ -161,343 +161,4 @@ func (d *dalalStreamService) getSubscription(req *datastreams_pb.SubscriptionId,
 	}
 
 	return subscription, nil
-}
-
-func (d *dalalStreamService) GetStockHistoryUpdates(req *datastreams_pb.SubscriptionId, stream pb.DalalStreamService_GetStockHistoryUpdatesServer) error {
-	var l = logger.WithFields(logrus.Fields{
-		"method":        "GetStockHistoryUpdates",
-		"param_session": fmt.Sprintf("%+v", stream.Context().Value("session")),
-		"param_req":     fmt.Sprintf("%+v", req),
-	})
-	l.Infof("GetStockHistoryUpdates requested")
-
-	subscription, err := d.getSubscription(req, datastreams_pb.DataStreamType_STOCK_HISTORY)
-	if err != nil {
-		return err
-	}
-
-	subscribeReq := subscription.subscribeReq
-	done := subscription.doneChan
-	updates := make(chan interface{})
-
-	stockId, _ := strconv.ParseUint(subscribeReq.DataStreamId, 10, 32)
-
-	historyStream := d.datastreamsManager.GetStockHistoryStream(uint32(stockId))
-	historyStream.AddListener(done, updates, req.Id)
-
-loop:
-	for {
-		select {
-		case <-done:
-			break loop
-		case <-stream.Context().Done():
-			d.removeSubscriptionFromMap(req)
-			close(done)
-			break loop
-		case update := <-updates:
-			err := stream.Send(update.(*datastreams_pb.StockHistoryUpdate))
-			if err != nil {
-				// log the error
-				break
-			}
-		}
-	}
-	l.Infof("Request completed successfully")
-
-	return nil
-}
-
-func (d *dalalStreamService) GetMarketDepthUpdates(req *datastreams_pb.SubscriptionId, stream pb.DalalStreamService_GetMarketDepthUpdatesServer) error {
-	var l = logger.WithFields(logrus.Fields{
-		"method":        "GetMarketDepthUpdates",
-		"param_session": fmt.Sprintf("%+v", stream.Context().Value("session")),
-		"param_req":     fmt.Sprintf("%+v", req),
-	})
-	l.Infof("GetMarketDepthUpdates requested")
-
-	subscription, err := d.getSubscription(req, datastreams_pb.DataStreamType_MARKET_DEPTH)
-	if err != nil {
-		return err
-	}
-
-	subscribeReq := subscription.subscribeReq
-	done := subscription.doneChan
-	updates := make(chan interface{})
-
-	stockId, _ := strconv.ParseUint(subscribeReq.DataStreamId, 10, 32)
-
-	depthStream := d.datastreamsManager.GetMarketDepthStream(uint32(stockId))
-	depthStream.AddListener(done, updates, req.Id)
-
-loop:
-	for {
-		select {
-		case <-done:
-			break loop
-		case <-stream.Context().Done():
-			d.removeSubscriptionFromMap(req)
-			close(done)
-			break loop
-		case update := <-updates:
-			err := stream.Send(update.(*datastreams_pb.MarketDepthUpdate))
-			if err != nil {
-				// log the error
-				break
-			}
-		}
-	}
-	l.Infof("Request completed successfully")
-
-	return nil
-}
-
-func (d *dalalStreamService) GetMarketEventUpdates(req *datastreams_pb.SubscriptionId, stream pb.DalalStreamService_GetMarketEventUpdatesServer) error {
-	var l = logger.WithFields(logrus.Fields{
-		"method":        "GetMarketEventUpdates",
-		"param_session": fmt.Sprintf("%+v", stream.Context().Value("session")),
-		"param_req":     fmt.Sprintf("%+v", req),
-	})
-	l.Infof("GetMarketEventUpdates requested")
-
-	subscription, err := d.getSubscription(req, datastreams_pb.DataStreamType_MARKET_EVENTS)
-	if err != nil {
-		return err
-	}
-
-	done := subscription.doneChan
-	updates := make(chan interface{})
-
-	marketEventsStream := d.datastreamsManager.GetMarketEventsStream()
-	marketEventsStream.AddListener(done, updates, req.Id)
-
-loop:
-	for {
-		select {
-		case <-done:
-			break loop
-		case <-stream.Context().Done():
-			d.removeSubscriptionFromMap(req)
-			close(done)
-			break loop
-		case update := <-updates:
-			err := stream.Send(update.(*datastreams_pb.MarketEventUpdate))
-			if err != nil {
-				// log the error
-				break
-			}
-		}
-	}
-	l.Infof("Request completed successfully")
-
-	return nil
-}
-
-func (d *dalalStreamService) GetMyOrderUpdates(req *datastreams_pb.SubscriptionId, stream pb.DalalStreamService_GetMyOrderUpdatesServer) error {
-	var l = logger.WithFields(logrus.Fields{
-		"method":        "GetMyOrderUpdates",
-		"param_session": fmt.Sprintf("%+v", stream.Context().Value("session")),
-		"param_req":     fmt.Sprintf("%+v", req),
-	})
-	l.Infof("GetMyOrderUpdates requested")
-
-	subscription, err := d.getSubscription(req, datastreams_pb.DataStreamType_MY_ORDERS)
-	if err != nil {
-		return err
-	}
-
-	done := subscription.doneChan
-	updates := make(chan interface{})
-
-	userId := getUserId(stream.Context())
-	myOrdersStream := d.datastreamsManager.GetMyOrdersStream()
-	myOrdersStream.AddListener(done, updates, userId, req.Id)
-
-loop:
-	for {
-		select {
-		case <-done:
-			break loop
-		case <-stream.Context().Done():
-			d.removeSubscriptionFromMap(req)
-			close(done)
-			break loop
-		case update := <-updates:
-			err := stream.Send(update.(*datastreams_pb.MyOrderUpdate))
-			if err != nil {
-				// log the error
-				break
-			}
-		}
-	}
-	l.Infof("Request completed successfully")
-
-	return nil
-}
-
-func (d *dalalStreamService) GetNotificationUpdates(req *datastreams_pb.SubscriptionId, stream pb.DalalStreamService_GetNotificationUpdatesServer) error {
-	var l = logger.WithFields(logrus.Fields{
-		"method":        "GetNotificationUpdates",
-		"param_session": fmt.Sprintf("%+v", stream.Context().Value("session")),
-		"param_req":     fmt.Sprintf("%+v", req),
-	})
-	l.Infof("GetNotificationUpdates requested")
-
-	subscription, err := d.getSubscription(req, datastreams_pb.DataStreamType_NOTIFICATIONS)
-	if err != nil {
-		return err
-	}
-
-	done := subscription.doneChan
-	updates := make(chan interface{})
-
-	userId := getUserId(stream.Context())
-	notificationsStream := d.datastreamsManager.GetNotificationsStream()
-	notificationsStream.AddListener(done, updates, userId, req.Id)
-
-loop:
-	for {
-		select {
-		case <-done:
-			break loop
-		case <-stream.Context().Done():
-			d.removeSubscriptionFromMap(req)
-			close(done)
-			break loop
-		case update := <-updates:
-			err := stream.Send(update.(*datastreams_pb.NotificationUpdate))
-			if err != nil {
-				// log the error
-				break
-			}
-		}
-	}
-	l.Infof("Request completed successfully")
-
-	return nil
-}
-
-func (d *dalalStreamService) GetStockExchangeUpdates(req *datastreams_pb.SubscriptionId, stream pb.DalalStreamService_GetStockExchangeUpdatesServer) error {
-	var l = logger.WithFields(logrus.Fields{
-		"method":        "GetStockExchangeUpdates",
-		"param_session": fmt.Sprintf("%+v", stream.Context().Value("session")),
-		"param_req":     fmt.Sprintf("%+v", req),
-	})
-	l.Infof("GetStockExchangeUpdates requested")
-
-	subscription, err := d.getSubscription(req, datastreams_pb.DataStreamType_STOCK_EXCHANGE)
-	if err != nil {
-		return err
-	}
-
-	done := subscription.doneChan
-	updates := make(chan interface{})
-
-	stockExchangeStream := d.datastreamsManager.GetStockExchangeStream()
-	stockExchangeStream.AddListener(done, updates, req.Id)
-
-loop:
-	for {
-		select {
-		case <-done:
-			break loop
-		case <-stream.Context().Done():
-			d.removeSubscriptionFromMap(req)
-			close(done)
-			break loop
-		case update := <-updates:
-			err := stream.Send(update.(*datastreams_pb.StockExchangeUpdate))
-			if err != nil {
-				// log the error
-				break
-			}
-		}
-	}
-
-	l.Infof("Request completed successfully")
-
-	return nil
-}
-
-func (d *dalalStreamService) GetStockPricesUpdates(req *datastreams_pb.SubscriptionId, stream pb.DalalStreamService_GetStockPricesUpdatesServer) error {
-	var l = logger.WithFields(logrus.Fields{
-		"method":        "GetStockPricesUpdates",
-		"param_session": fmt.Sprintf("%+v", stream.Context().Value("session")),
-		"param_req":     fmt.Sprintf("%+v", req),
-	})
-	l.Infof("GetStockPricesUpdates requested")
-
-	subscription, err := d.getSubscription(req, datastreams_pb.DataStreamType_STOCK_PRICES)
-	if err != nil {
-		return err
-	}
-
-	done := subscription.doneChan
-	updates := make(chan interface{})
-
-	stockPricesStream := d.datastreamsManager.GetStockPricesStream()
-	stockPricesStream.AddListener(done, updates, req.Id)
-
-loop:
-	for {
-		select {
-		case <-done:
-			break loop
-		case <-stream.Context().Done():
-			d.removeSubscriptionFromMap(req)
-			close(done)
-			break loop
-		case update := <-updates:
-			err := stream.Send(update.(*datastreams_pb.StockPricesUpdate))
-			if err != nil {
-				// log the error
-				break
-			}
-		}
-	}
-
-	l.Infof("Request completed successfully")
-
-	return nil
-}
-
-func (d *dalalStreamService) GetTransactionUpdates(req *datastreams_pb.SubscriptionId, stream pb.DalalStreamService_GetTransactionUpdatesServer) error {
-	var l = logger.WithFields(logrus.Fields{
-		"method":        "GetTransactionsUpdates",
-		"param_session": fmt.Sprintf("%+v", stream.Context().Value("session")),
-		"param_req":     fmt.Sprintf("%+v", req),
-	})
-	l.Infof("GetTransactionsUpdates requested")
-
-	subscription, err := d.getSubscription(req, datastreams_pb.DataStreamType_TRANSACTIONS)
-	if err != nil {
-		return err
-	}
-
-	done := subscription.doneChan
-	updates := make(chan interface{})
-
-	userId := getUserId(stream.Context())
-	transactionsStream := d.datastreamsManager.GetTransactionsStream()
-	transactionsStream.AddListener(done, updates, userId, req.Id)
-
-loop:
-	for {
-		select {
-		case <-done:
-			break loop
-		case <-stream.Context().Done():
-			d.removeSubscriptionFromMap(req)
-			close(done)
-			break loop
-		case update := <-updates:
-			err := stream.Send(update.(*datastreams_pb.TransactionUpdate))
-			if err != nil {
-				// log the error
-				break
-			}
-		}
-	}
-	l.Infof("Request completed successfully")
-
-	return nil
 }
